requirements.txt
numpy==1.21.6
opencv_python==4.6.0.66
pandas==1.3.5
Pillow==9.4.0
tensorflow==2.9.2
torchvision==0.14.1+cu116
